Meta Generated Items renderer overhaul
