+++ src/main/java/gregtech/api/interfaces/IRegularMaterialRenderer.java
@@ -1,0 +1,10 @@
+package gregtech.api.interfaces;
+
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.IIcon;
+import net.minecraftforge.client.IItemRenderer;
+
+public interface IRegularMaterialRenderer {
+
+    void renderRegularItem(IItemRenderer.ItemRenderType type, ItemStack aStack, IIcon icon, IIcon overlay);
+}
